Release v3.36.0

+++ inc/taxonomies-functions.php
@@ -1977,16 +1977,29 @@ function taxopress_unregister_taxonomy($taxonomy)
 
     global $wp_taxonomies;
 
-    $taxonomy_object->remove_rewrite_rules();
-    $taxonomy_object->remove_hooks();
+    // Instead of removing totally, keep minimal structure but disable functionality
+   if (isset($wp_taxonomies[$taxonomy])) {
+        $wp_taxonomies[$taxonomy]->public = false;
+        $wp_taxonomies[$taxonomy]->show_ui = false;
+        $wp_taxonomies[$taxonomy]->show_in_menu = false;
+        $wp_taxonomies[$taxonomy]->show_in_nav_menus = false;
+        $wp_taxonomies[$taxonomy]->show_in_rest = false;
+        $wp_taxonomies[$taxonomy]->show_tagcloud = false;
+        $wp_taxonomies[$taxonomy]->show_in_quick_edit = false;
+        $wp_taxonomies[$taxonomy]->show_admin_column = false;
+        $wp_taxonomies[$taxonomy]->query_var = false;
+        
+        // Remove rewrite rules
+        if ($taxonomy_object) {
+            $taxonomy_object->remove_rewrite_rules();
+            $taxonomy_object->remove_hooks();
+        }
 
-    // Remove custom taxonomy default term option.
-    if (!empty($taxonomy_object->default_term)) {
-        delete_option('default_term_' . $taxonomy_object->name);
+        // Remove custom taxonomy default term option.
+        if (!empty($taxonomy_object->default_term)) {
+            delete_option('default_term_' . $taxonomy_object->name);
+        }
     }
-
-    // Remove the taxonomy.
-    unset($wp_taxonomies[$taxonomy]);
 
     /**
      * Fires after a taxonomy is unregistered.
